requirements.txt
streamlit
openai-whisper
ollama
fpdf
